Make build results more responsive friendly
Remove unneeded marging bottom mb-2 for build results. This is already
controlled by the parent card-body element.
Remove unneeded horizontal padding px-3, that wastes space in smaller
screens.
Add no-gutters class to have better control of paddings in row and col
elements.
Define left paddings depending on the size of the display, on repos,
architectures and statuses. Useful for smaller displays.
Define and adjust col grid widths depending on the size of the display on
architectures and statuses. This fixes #8101.