Move binary specs to their own file
The Webui::PackageController spec is over 1600 lines long, it accurately
reflects what is in Webui::PackageController but we do not have to mimic the
length of the Class in the spec. Let's move out some specs for actions that
belong together so that at least the spec becomes a reasonable, manageable size.