Use staging_project_name instead of name
In the show route of staging_projects it was used `name` as the
parameter `name` of a staging_project. We change it to `staging_project_name`
for clarity, and because it is used this way in the rest of the routes.

Tests were also updated, as change of the review statuses require a user
to be logged.

Co-authored-by: David Kang <dkang@suse.com>