Don't loop over partial in staging project show history
Looping over partials is very inefficient and causes longer
rendering times. Moving the logic towards rendering
a collection is a bit difficult and the partials are only
used at one place. So just getting rid of the partials and
moving the code to the history template comes without any
disadvantages.

Related to #7350