Adjust Procfile to work with real-time indices
When `ts:rebuild` is called in the development environment to start the
Sphinx daemon, it never reaches `ts:rt:index` because `ts:start` runs with
`NODETACH=true` and takes the control of the console, so it never returns.

The hack was to add code in Procfile to run in paralell a process that
after a while (time needed for the Sphinx daemon to start) builds the
real-time indices.

Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>