Switch to Thinking::Sphinx real-time indices
Instead of reindexing every hour with SQL-backed indices, we now use
real-time indices. That allows us to get results when searching by
just created or updated projects or packages.

Real-time indices require references to models' methods when defining attributes.
For this reason, it was needed to add methods like `attribs_attrib_type_ids`,
`packages_package_issues_ids` or `devel_packages?` for both package and
project indices.

Real-time indices are updated with callbacks, so we don't need an external
process for updating indices. Temporarily, the entry in the `config/clock.rb`
file for updating the indices is commented.

The search can also be done by attributes and issues. Therefore, when
attributes or issues are created/updated, an update of the indices
is done through callbacks. Also, every time an issue of a package
is updated, both indices of the package and its related project are
updated.

All the other attributes of the indices are used for sorting the
results. For now the indices are not going to be updated when the
values of these attributes change.

It was fixed the way the `activity_index` is obtained for a project.
Before it took the maximum `activity_index` value of its packages that had
the oldest `updated_at` value. Now it takes the maximum `activity_index`
value of its packages that have the more recent `updated_at` value.

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>