Create Sphinx start Rails task for development...
... environment.

Provide a way to reindex Sphinx when it is started in NODETACHED mode.
This is only needed in development environment.

The check for rebuilding the configuration of the index was modified to
include the condition of no existence of the *.sphinx.conf file. That
is the case of the development environment.

This is a workaround until pat/thinking-sphinx#1150 is fixed.