Improve error message for non-existent staging project
Before, if the staging project didn't exist, it raised an exception with a quite cryptic error message.
Now, it shows a descriptive error message.

It uses a new exception defined in the Source::Errors module instead of
raising a generic exception with a customized message.

The show and accept actions had some code in common, it has been moved
to a before_action.