Prevent an exception to be raised in auto cleanup
The performance of the auto cleanup delayed job failed because the check
of dependencies raised an exception. So the delayed job was retried
again and again.

The exception was caused by a project with a develpackage inside, that
could not be removed.

Now, instead of raising an exception, the dependencies are correctly
checked.

Fixes: #8762

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>