start storing build stats whenever possible.
THe stats are stored in the simple file format, so external processsing
can be easy. This allows changes to the format and structure. The
xml from _statistics file parsed into an hash and is flattened into an
array. This xml template used to parse the _statistics file is also
flattened into an array. This array acts as an overlay to store the
stats.
Since the same xml template is used to parse xml and as layout of
storage, this patch acts like a converter from xml to tabular format.

A simple and implementation independent file format is chosen to allow
easy enhancement to the schema without affecting the application code