[backend] generate _statistics file when the result is no-change
this patch moves some of the code around into a new function . It adds a
few lines for the real work of creating file when there is no change.
The code for the new function `updatestatsfromfile` add stats to the
passed in hash after reading it from file. The function is not
responsible for checking the existence of file and its format.