[dist] Update rake version to 12.3.3 in bundled gems spec
When the selenium-webdriver gem was updated in 80e6a5d15c271a20c529b5c187a9090a023bc598,
rake was updated from 12.3.2 to 12.3.3. However, the obs-bundled-gems
RPM spec was not updated along with the rake update in the Gemfile.lock
to reference the newer rake version, so package builds began to fail.

This change fixes that by updating the %rake_version macro to the
correct version in the obs-bundled-gems RPM spec file.