[backend] bs_dispatch: speed up nativeonly handling
The nativeonly flag was checked in the body of two nested loops,
taking way too much time if there are many jobs and many idle
workers. Add a cache so that we can prune workers that we know
cannot build the job.

Also remove the worker from the idle list if its file is gone.
Not removing it led to lots of unneccessary constrain checks.