Authorize against the linked project instead of the package's project
When we're triggering the build from a linked project, the Pundit Policy checker
for the package fetches the package's project, not the one we're rebuilding
from.

This PR detects if we're triggering the build from a linked project and then
passes that project to Pundit to perform the authorization.

Fixes #6289

Co-authored-by: Henne Vogelsang <hvogel@opensuse.org>