Fixed logout redirection when anonymous access is not allowed.
If anonymous access is not allowed and a user logs out the
user is redirected back to the refering page prior to logout.  This
redirection produces a error since anything other than the root path
is now unavailable to the user.

This patch will redirect the use to the root path on logout if anonymous
access is not allowed.