Move the GPG Key and SSL Certificate details to its own page
The GPG Key and SSL Certificate details was present in a modal embedded into
the Project overview. On small devices it was harder to see all the data as you
needed to scroll to the right.

Extracting the details into a new page provides more space. We also decided not
to use a table to present the data to make it easier to select and copy the
interesting bits.