Fix missing public key from project directory.
Problem: When a global signing key is defined and forceprojectkeys is not
set, the default public key defined in BSConfig::keyfile is not placed
in the project/projid.pkg directory.  This creates a problem for local
builds.  Packages are signed using the default key. When the packages
are used during local builds the verification fails because there
isn't a public key to validate the packages from that project.

The default public key is correctly added to the repository during
publishing.

This patch mimics the process the publish routine uses to place the default
pubkey file during project creation.