Don't overwrite projid.pkg/_pubkey with the default if one already exists.
One scenerio where this may exists is when changing DOD projects.
Without this addtional check, if a project with a DOD project with an
existing _pubkey file to match the dod packages, the _pubkey will
be replaced with the default key.  It is simply safer to not
replace the _pubkey if it already exists.