Compare paths stricter in tab_link helper
Using 'include?' to compare the paths (string class) leads
to problems, since the tab items got marked as active
if only parts of the paths are matching. For those routes
an additional variable got passed to the method in order to
prevent those side effects. By comparing the paths with
the eql operator, the helper method loses complexity, since
one of the named parameters can be dropped.