Implement in-place editing for some of the project's fields
Use in-place editing for the project's
- Title
- Url
- Description

We've been using modals to edit the project's title, url and description. We
think we have better tools for that. In-place editing is one of the ways to do
it.

With this PR we introduce a new way of updating some of the project's
information from the projects overview page.

Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>