Add in-place editing for project's basic info
The project's basic information (title, URL and description) is now
editable in-place.

By clicking the "Edit Project" link inside the Actions menu, the static
content is replaced by a form that allows to modify the project's basic
information.

An animation shows a smoothly transition between static content and
form.

UJS is used for all the action: edit, update and back to show (on
Cancel) .

The Ajax form's validation is handled manually.

Co-authored-by: Daniel Donisa <daniel.donisa@suse.com>