Add request state field to notifications table
At the time we access the state, it could already have been changed. We
need to store this field when the notification is created.

We store the "state" only when it is needed, for notifications coming
from events with type "RequestStatechange". The other request and comment
notifications don't make use of this "state" field.