[backend] BSSSL: fix potential file descriptor leak
The leak was introduced when we modernized the file handle
usage in commit 30b5fa2a25bcacf237c9a2ec61b97dde2f182585.
This led to a cyclic reference which meant that https
sockets were no longer closed automatically.

We now use 'local' again in BSSSL as a workaround. I hope that
this works with all perl versions.