[backend] Make HTTPS communication between services possible
When workers and services are located in different networks, we don’t
want to send traffic unencrypted and open to anyone to listen to.

Make it possible to easily opt in to HTTPS communication by changing
the service URLs in BSConfig.

This is based on a pull request created by Andrej Shadura
<andrew.shadura@collabora.co.uk>. Thanks a lot!