Change CircleCI config to use new ci docker repo
The new ci docker containers project has been created with the idea of
creating the containers used in the CI test suite for Unstable. So
upgrading the version of the underlying system of the containers doesn't
affect the CirleCI configuration any more.

This new ci docker containers repo is now using SLE15 SP1 docker base
images, instead of SLE12 SP4.