[backend] findpackages: move multibuild processing after project link processing
We used to do multibuild processing first and relied on the project link
code to return the multibuild packages from the link. The problem with
this is that this uses the wrong multibuild data.

This commit changes findpackages_projlink to no longer return multibuild
packages and moves the addmultibuildpackages call after the
findpackages_projlink call.