Refactor code related to events creation
Before, when a change in an object triggered an event creation, the method
'create_notification' was called. It calculated the set of parameters needed to
create the event (according to the type) and then created the event.

Now, the same functionality is splitted in two methods: 'event_parameters'
(public) and 'create_event' (private). This way the set of parameters is
accesible from other parts of the code.

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>