Get the parameters for notifications from each model
When creating a notification we use a set of parameters. Some of them
are common (event type, payload, etc.) and some others depend on the
type of event.

Before, we calculated all the possible sets of parameters in a concrete
point of the code by checking the event type. Now, we added a method
called 'parameters_for_notification' to each kind of event so we just
ask them for its set of parameters.

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>