Define NotificationCreator service
The notification creation code used to be inside the SendEventEmailsJob.
The parameters needed to create the notifications were also calculated
there (according to the type of event).

Now, the creation is isolated in the NotificationCreator service, which
keeps the code tidy and makes it accesible from other parts of the code.

The calculation of the parameters has been distributed to the
corresponding models. They all have now a method called
'parameters_for_notification' to do so.

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>