Add data migration to remove obsolete notifications
Recent refactoring on notification creation code and database structure
left some inconsistent data in the database.

To solve the problem, with the current data migration we are going to
remove the affected notifications as first step. Then, they will be
re-generated in the following data migration.

Data migration's tests included.

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>