Track date for all the notifiable events
A few event types used to track, in its payload, the date and time when
the event was triggered. The 'when' key was used for this. From now on,
all the notifiable event types track this value.

Now, when a notification is created, its timestaps fields (create_at and
updated_at) are set to the 'when' value. This way, we ensure they have
realistic date and time according to the moment when the event happened
and not when the notification was created (what usually happens some
time later).