Allow showing the notifications all at once
Add a link to show all the notifications in the same page.
There is a limit, MAX_PER_PAGE. In case of exceeding this number,
a flash message will be shown and the results will be paginated
with MAX_PER_PAGE notifications per page.

This makes it easier for users to search texts on the list.