Take file name into account for file partial caching
In certain cases, the md5 and mtime value of a file is the same
for multiple files. Only the file name differs in those cases.
This needs to be considered, in order to have uniq data in the
cache.

In the past this lead to not show certain files in the
webui due to this https://github.com/rails/rails/pull/35145
Since this got fixed, it now does the opposite (showing duplicated entries),
which is more obvious.

Since osc shows the correct files, and missing files are less
obvious than duplicates, this probably took a while to be reported
as an issue.

Fixes #9352

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>