Create notified_projects join table
Retrieving the projects of a notification, requires querying for comments,
requests and reviews. We want to avoid n+1 queries instead of going through
the polymorphic associations notifiable and notifiable.commentable for
comments.

We need to have an ID in the table since we'll be using a `has_many through:`
association

Co-authored-by: Dany Marcoux <dmarcoux@suse.com>
Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>