Use back notifiable to get the corresponding object
For a while, the notifiable_id values were wrong in the database,
so we couldn't rely on it to get the "notifiable" object (review,
comment) and its request.

As a temporary solution, we took the request number from the event
payload.

Now that the values in the database are correct (after a data migration)
the workaround is reverted.