Use Notification instead of Notification::RssFeedItem
We plan to remove the column type from `Notification` and remove the
class `Notification::RssFeedItem` and only use Notification class.

The first step is not use this class anymore. After that we should
create a migration to remove the column type and get rid the class
`Notification::RssFeedItem`.

For that we created new `scopes` to fetch the notifications for the
different channels.