Remove notifications without notifiable
Previously, we excluded the events `Event::BuildFail` and
`Event::ServiceFail` for creating rss and web notifications.
We also regenerated the notifications for the users except for
those with event_type `Event::BuildFail` and `Event::ServiceFail`.

This commit, will remove the notifications without notfiable, fixing
at the same time the inconsistency in the DB and also the notifications
web counter.