Refactor repositories view
For beta users, the general actions are shown in Actions menu.
Therefore, the first section/card of the repositories page keeps empty.
In that case, the view is slightly modified and the two first
sections/cards are merged.

To do so, some new conditions are added to the code, what increases
its complexity. Therefore, the view code has been refactored.