Hide build results for delete requests
When we are in a request for package deletion, it makes no sense to
build the package which no longer exists.

Using 'defined?' to ckeck if the project or package exist is not a good
idea because it always returns truthy (returns "local-variable") even if they are nil.
Therefore, 'defined?' has been replaced by '||='.