Do not blindly read an uploaded file into memory
Currently, when uploading a file, its entire content is read into
memory and then passed to Package.verify_file! (even if the
PackageService::FileVerifier, which is called by Package.verify_file!,
never uses the content). Instead of always reading the file into
memory, it should be only read if its content is really needed in the
PackageService::FileVerifier class.
For this, we now pass a "readable" object to Package.verify_file!.
The "readable" object defers the call to request.raw_post until
its "read" method is called. The read method is only called if a
service class really needs the content. Moreover, the read method
is at most called once by the PackageService::FileVerifier class,
which "caches" the result.

Note: it is still the case that large files are read into memory.