Always verify an uploaded file in case of a _pattern package
As of now, if a file is PUTed to a _pattern package, the file is
only verified if the _pattern package was non-existent. Instead, the
the file should be always verified (regardless whether the _pattern
package was created or already existed).
Since @pack is not set in check_permissions_for_file in case of a
_pattern package (which is perfectly fine), it is set up in
update_file (the code also takes a potential race into account).
Hence, @pack is always defined when it is passed to
Package.verify_file!.
The new behavior is documented in a testcase.