[backend] limit reading jobhistory to 1000 entries
jobhistory can be really large (imagine all of project history, since
the beginning of time). If some package rebuild failed and was never
rebuilt(successfully) since ages(searching for that package needs entire
jobhistory(which could be in gbs) to be read) we will have huge memory
consumption.

In otherwords, package build failed now and a prior rebuild was
Gigabytes ago (or not all) then we will have memory issues.