[backend] maintain the history of the last failed jobs
We write a .failhistory file that contains the last failed jobs for
a package. If a build succeedes, the .failhistry is deleted again.

We use this as a cheap way to get the number of successive failures
when sending notifications.
It is also useful when trying to find problems with some of the workers.