Add communication scope
Communication can be displayed for a specific group of users, what we call communication scope.
At the moment the groups are:
  - All users
  - Logged in users
  - Admin users
  - In beta users
  - In rollout users

"All users" is the default communication scope, so this option is made as
default in database.

Part of the code is temporarily duplicated in StatusMessage and
Announcement until both are merged in the future.