Add communication scope to status messages
Certain status messages can be displayed for a specific group of users,
what we call communication scope.

At the moment the groups are:
  - All users
  - Logged in users
  - Admin users
  - In beta users
  - In rollout users

"All users" is the default communication scope. To avoid downtime when
setting this default value in database, two migrations has been added.
One to add the field and another to backfill it with the default value
in all the existent records.

Co-authored-by: Victor Pereira <vpereira@suse.com>