Move 'Submit Package' modal to its own page
There's a new controller which uses strong parameters, form builders and
nested attributes for the bs_request_actions association of the
BsRequest model.

We rely on request_controller's create action instead of reimplementing
our own. The errors previously rescued in the action submit_request are
all inheriting from APIError, so they will still be caught by the
rescue.  As for superseding requests, we can do this in a filter after
the create action.

Co-authored-by: Dany Marcoux <dmarcoux@suse.com>
Co-authored-by: Daniel Donisa <daniel.donisa@suse.com>