Add announcement severity to StatusMessage
Announcement and StatusMessage functionalities are going to be merged,
they are quite similar.

To do so, StatuMessage's severity is going to include a new value:
"announcement". This kind of status message is going to be displayed in
a different way and need acceptance from the user.

An enum is defined for severity values and they will be introduced as
strings and not like integers.

Co-authored-by: Victor Pereira <vpereira@suse.com>