[api] support request delegation to package origin project.
Adds an option to let the server delegate the request for
a target project to the project where the packages is actual
living.

This works independend of a source package link, eg from submitting
from a base distro project to another one.

It is needed for Jump submissions in first place (jsc#OBS-59)