Add communication scope to status messages
Certain status messages can be displayed for a specific group of users,
what we call communication scope.

At the moment the groups are:
  - All users
  - Logged in users
  - Admin users
  - In beta users
  - In rollout users

"All users" is the default communication scope. To avoid downtime when
setting this default value in database, two migrations has been added.
One to add the field and another to backfill it with the default value
in all the existent records.

If the status message was created with "announcement" severity,
it is displayed on the top of the page.

The displayed announcement will always be the newest one among those the
user is allowed to see. In case that announcement is already
acknowledged by the user, it will not be displayed.

The StatusMessage new for has been moved from modal to its dedicated new
page.

Co-authored-by: Victor Pereira <vpereira@suse.com>

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Victor Pereira <vpereira@suse.com>