[backend] Support a source copy without creating a commit
By passing the "rev=repository" query parameter to the "copy"
route, the source files of the origin package are just made available
in the target package (for instance, for later processing). No commit
to the target package is made during this operation (see the addrev
with "rev=repository" semantics).
For a potential use case, see [1].

[1] https://github.com/openSUSE/open-build-service/issues/9615