Modify data migration for announcements
The data migration to dump announcements into status messages has been
modified to make it work even after the Announcement model is removed
(something that is planned for the near future).

Inside the data migration file, we define "fake" models to use the
tables involved. The tables will be removed, but this data migrations is
going to be executed first.

Co-authored-by: David Kang <dkang@suse.com>