Move files-related actions to their own controller
This commit is the first of a series of commits to move all the
actions related to `file` into `FilesController`.

In the previous action `add_file` there was a method call
`set_file_details` that was not used at all in this action. This method
instanctiate some variables that are used on `show` view but not in `add_file`.