Move modules and classes related to backend communication to app/lib
The migration towards zeitwerk requires little adjustments on the
file tree of the backend modules in order to follow the conventions.
Zeitwerk obtains the constant names from the filenames
and considers folder nesting. Therfore we would need a backend.rb file
in the root of the lib folder and then to eager load the whole
lib folder.

The better option is to move the backend module to the app folder,
since everything in there get eager-loaded/auto-loaded by default.
This is the new recommended and modern idiom.