Install annotate gem and configure auto_annotate_models task
Annotate was dropped in the past because of some conflicting executables:
289da9399fb7116229bc6972d5800442062c6b57

Now we can bring it back, so the last version of the gem is installed.

The task lib/tasks/auto_annotate_models.rake update the annotation on
every db:migrate. The configuration in the file has been modified to fit
our needs.