Increase exclude-limit to 100 for RuboCop TODOs
With the default value of 15 for `exclude-limit`, whenever a cop
registers offenses for more than 15 files, the cop is disabled in the
TODOs file. With this commit's changes, this limit is raised from 15 to
100. This makes it easier to tackle TODOs one file at a time since we
know which files are problematic. Only cops with registered offenses in
more than 100 files are still disabled in the TODOs file.

TODOs were regenerated to take this change into account.