Correct class name of group maintainers db migration
The database migration for the group maintainers used the
same class name as the corresponding model. This leads
to conflicts and throws an 'superclass missmatch' exception
since the two classes inherit from different classes.

This problem got triggered by the move to the zeitwerk
codeloading engine, but was wrong already before.