Enable cache versioning
This commit doesn't change the current behaviour of the cache, but
enables cache versioning either for future use cases, or to change current
cases.

In this commit we adapt the existing code substituting the cache key methods.

Before cache versioning, `cache_key` contained the key and the version.

Introducing cache versioning makes `cache_key` to contain only the key.
In order to have the same content as before, we have to use
`cache_key_with_version`.