Switch smoke tests to apparition capybara driver
This driver is being developed by the current developer of
Capybara and will attempt to keep up to date with new Capybara releases.
We expect this to be more stable.

Esecially as it uses CDP to communicate with Chrome, thereby obviating
the need for chromedriver.

The capybara gem had to be updated to support apparition.