Regenerate RuboCop TODOs
With the introduction of schema.rb, some RuboCop cops are now active and
linting the schema for potential offenses. One of them
(Rails/UniqueValidationWithoutIndex) is registering offenses and to keep
changes as small as possible, we don't want to address the offenses in
this PR. We could have disabled the cops, but they do provide value. In
the end, regenerating the RuboCop TODOs makes the CI pass and allows us
to tackle those offenses in a separate PR.