Create data migration to adapt review and duplicated notifications
Instead of having dedicated notification for review wanted events,
we will from now on unify this with the request notifications.
The association to the review will be dropped. Therefore we
need to associate the review notifications to the corresponding
request through a data migration.
Also, we are going to show only the latest state of a request
or comment through the notifications. The data migration
needs to remove duplicated/outdated notifications that already
exist.

Co-authored-by: David Kang dkang@suse.com