Clean up old or irrelevant notifications
We delete old or outdated notifications for Requests and Comments.
When someone adds a Review the notifications page will show a
notification for a Request instead.

The notifications page can get a bit crowded with notifications
sometimes.
Most of those notifications would be no longer relevant because the
latest will have the most up to date information. With this change the
notifications page will only show useful and up to date data about
your Requests.

Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Lukas Krause <lkrause@suse.de>