Require auth for binary action
Displaying information about a binary is a rather expensive backend
action that we should not allow for anonymous users.

We just had a case of a (bad) spider taking too many resources by
querying all binaries of a distribution. If someone needs this
information they can get it from the repository/binary data on the
mirrors.