Prevent exception on review wanted rss feeds
The recent changes on the notification system, included
dropping the creation of dedicated notifications of type
review for 'ReviewWanted' events. Instead they are also
associated to the corresponding BsRequst.
This is not handled by the new RSS feed templates, which leads
to an exception. To prevent the exception on ReviewWanted events,
we can switch back to the older version of the templates, which
base on the event payload itself. Since the event payload
hasn't changed, the older templated are still working as expected.

Fixes #9899

Co-authored-by: Daniel Donisa <daniel.donisa@suse.com>
Co-authored-by: David Kang <dkang@suse.com>