[backend] user/pass for obs services called in docker
This patch adds the possibilty to specify a user/password combination
in BSConfig.pm and export to ENV in the container to use for basic auth
at the API server (specified in api_url).
This is required e.g. for kiwi_import.

Without this patch the service fails with '401 - Unauthorized' when using
e.g. 'https://build.opensuse.org' as api url.