[backend] docker network for services configurable
Some of our services like bundle_gems use a local proxy/cache like geminabox.
So far we used the "--link" option to make the geminabox container accessible
by name for the service container.

Using "--link" on a bridge is deprecated and only worked so far because we
patched the docker package on the service server.

This patch makes the network name configurable so we can use user-defined
networks in the future.