Adapt thinking sphinx callback syntax
Thinking Sphinx v5.0 introduces explicit callbacks. Since we
are using the real time indices, we already have explicit
callbacks set. But at the same time, a new syntax for the
callback got introduced with:
-> https://github.com/pat/thinking-sphinx/pull/1175
The old way is still supported, but in the changelogs it
is recommended to adapt the new style right away.