Remove session tracking
This was an attempt to measure logins with proxy authentication based
on User.session. But User.session is a thread variable that is reset
after each request. So it's basically impossible to check based on this.
The user migh just have authenticated or we might just be on a new
thread ¯\_(ツ)_/¯

It also doesn't make much sense to measure :unauthenticated and
:disabled without measuring the rest of the case.

So let's stop measuring this at all until we find a better way.

For non-proxy auth this is still measured in Webui::SessionController.