Explicitly send the notifiable type to the partials
Instead of picking the notifiable type from the `params` variable,
this PR makes the notifiable type explicitly available as a partial
local variable and function parameters.

Before this PR, the user was losing the selected notifiable type from the
notifications filter when marking a notification as 'read'.

This was due to the notifications/update remote partial not getting
the `params[:type]` when rendering.

Fixes #9908