Show download link for files that exceed 4 megabytes
In the past we showed the download link for files with
a size of less than 4 megabytes also to non authenticated users.
Now we show the link only to authenticated users no matter which
sizes the file has. A forgotten if statement led to the problem
that we didn't showed the download link to anyone anymore when
the file exceeds the 4 megabytes.

Fixes #10096