Implement new z-index hierarchy
We have issues with components, overlapping each other in the wrong way.

This pull request changes the z-index hierarchy of the navigation bars and
side panels, using default values provided by Bootstrap, and removing z-index
changes that were considered unnecessary.

In the case of the old UI, only the watchlist panel was changed, to use
a z-index lower than the one used by default in Bootstrap for modals.

In the case of the new responsive UI:
- fixed elements (fixed-top and fixed-bottom) in the top navigation bar
  and the bottom navigation bar were overwritten to a z-index lower than
  the one used by default in Bootstrap for modals.
- the watchlist panel z-index has a even lower value, to allow to show
  at the same time the user dropdown, located at the upper-right corner.

The value of a z-index lower than the one used by default in Bootstrap's
modals was chosen because we use modals for the login and sign-up forms.
We want these modals to be shown over the top and bottom navigation bars.