Only use GET for search index endpoint
The search#index endpoint doesnt modify or
store and any data. Therefore we shouldn't allow
or use POST request to access the resource.
This will also prevent issues with our 404 page
which is a static html file that doesn't include a
CSRF token (which would be require for POST).

Fixes #10196