Update BS request with AJAX call
The commit introduces asynchronous update for BS Requests.

This was implemented in scope of adding notification toolbar feature
(#10220).
Previously, when the request got updated (e.g. approved/rejected), the
page was refreshed and notification toolbar was removed, because it
relies on 'notification_id' query parameter, but that parameter got
removed after action was made on BS request.